refactor: 优化 DictManager,拆分为 DictWriter 和 DictReader
