+++ Fire/Dict/DictReader.swift
@@ -1,74 +1,26 @@
 //
-//  DictManager.swift
+//  DictReader.swift
 //  Fire
 //
-//  Created by 虚幻 on 2022/7/2.
-//  Copyright © 2022 qwertyyb. All rights reserved.
+//  Extracted from DictManager.swift: read-only dictionary query logic.
 //
 
 import Foundation
 import Defaults
 
-class DictManager {
-    static let shared = DictManager()
-    static let userDictUpdated = Notification.Name("DictManager.userDictUpdated")
-
-    let tempEnTriggerPunctuation: Character = ";"
-
-    /// 解析用户词库行，支持双引号包裹含空格的词组
-    /// 例: "abc \"hello world\" test" → ["abc", "hello world", "test"]
-    private func parseQuoteAware(line: Substring) -> [Substring] {
-        var result: [Substring] = []
-        var current = line[...]
-        while !current.isEmpty {
-            current = current.drop(while: { $0.isWhitespace })
-            guard !current.isEmpty else { break }
-            if current.first == "\"" {
-                current = current.dropFirst()
-                let end = current.firstIndex(of: "\"") ?? current.endIndex
-                result.append(current[..<end])
-                current = current[end...]
-                if current.first == "\"" { current = current.dropFirst() }
-            } else {
-                let end = current.firstIndex(where: { $0.isWhitespace }) ?? current.endIndex
-                result.append(current[..<end])
-                current = current[end...]
-            }
-        }
-        return result
-    }
-
-    lazy var userDictFilePath: String = {
-        let dirs = NSSearchPathForDirectoriesInDomains(.applicationSupportDirectory, .userDomainMask, true)
-        guard let dir = dirs.first, let bundleID = Bundle.main.bundleIdentifier else {
-            return NSHomeDirectory() + "/Library/Application Support/Fire/user-dict.txt"
-        }
-        return dir + "/" + bundleID + "/user-dict.txt"
-    }()
-
-    /// 数据库指针（供 build.swift 迁移数据时读取）
-    private(set) var database: OpaquePointer?
+final class DictReader {
+    private var database: OpaquePointer?
     private var queryStatement: OpaquePointer?
 
-    private init() {
-        // 监听编码模式、候选词数、生僻字开关等偏好变更，自动更新 SQL 查询语句
-        Defaults.observe(keys: .codeMode, .candidateCount, .enableGBK, .enableEmoji) { () in
-            self.prepareStatement()
+    func bind(database: OpaquePointer?) {
+        self.database = database
+    }
+
+    func teardown() {
+        if queryStatement != nil {
+            sqlite3_finalize(queryStatement)
+            queryStatement = nil
         }
-        .tieToLifetime(of: self)
-    }
-    deinit {
-        close()
-    }
-    func reinit() {
-        close()
-        prepareStatement()
-    }
-    func close() {
-        queryStatement = nil
-        sqlite3_close_v2(database)
-        sqlite3_shutdown()
-        database = nil
     }
 
     private func getStatementSql() -> String {
@@ -123,45 +75,10 @@ class DictManager {
         """
         return sql
     }
-    
-    private func enableSQLiteProfile(_ db: OpaquePointer?) {
-        guard let db = db else { return }
-        
-        // 注册 profile 回调
-        sqlite3_profile(db, { _, sql, nanoseconds in
-            /*
-             * SQLite 内部计时精度 = 微秒级 / 毫秒级，不是纳秒级！
-             * 虽然 sqlite3_profile 给你的单位是 纳秒（UInt64），
-             * 但 SQLite 内部真正计时精度并没有那么高。
-             * 它的时间来源是：
-             * Windows：GetTickCount 精度 1ms ~ 16ms
-             * macOS / iOS：mach_absolute_time 转成后对齐到毫秒级别
-             */
-            let ms = Double(nanoseconds) / 1_000_000
-            let sqlStr = String(cString: sql!)
-            
-            FireLog.dict.debug("SQL duration: \(ms, privacy: .public)ms, sql: \(sqlStr)")
-        }, nil)
-    }
 
-    private func prepareStatement() {
-        if database == nil {
-            let rc = sqlite3_open_v2(getDatabaseURL().path, &database, SQLITE_OPEN_READWRITE, nil)
-            guard rc == SQLITE_OK else {
-                FireLog.dict.error("Failed to open database: \(String(cString: sqlite3_errmsg(self.database)), privacy: .public)")
-                return
-            }
-#if DEBUG
-            enableSQLiteProfile(database)
-#endif
-            sqlite3_exec(database, "PRAGMA case_sensitive_like=ON;", nil, nil, nil)
-            // 限制 SQLite 页缓存为 2MB，防止候选词查询缓存持续增长占用过多内存
-            sqlite3_exec(database, "PRAGMA cache_size=-2000;", nil, nil, nil)
-        }
-        if queryStatement != nil {
-            sqlite3_finalize(queryStatement)
-            queryStatement = nil
-        }
+    func prepare() {
+        teardown()
+        guard database != nil else { return }
         let sql = getStatementSql()
         // 日志输出实际执行的 SQL，方便调试查询问题
 #if DEBUG
@@ -174,15 +91,6 @@ class DictManager {
         }
     }
 
-    private func getMinIdFromDictTable() -> Int {
-        let sql = "select min(id) from wb_py_dict"
-        var minId: Int32 = 0
-        sqliteQuery(database, sql) { stmt in
-            minId = sqlite3_column_int(stmt, 0)
-        }
-        return Int(minId)
-    }
-
     // replaceTextWithVars 中使用的 DateFormatter 提为 static let，避免重复创建
     private static let dateFormatter: DateFormatter = {
         let f = DateFormatter()
@@ -192,7 +100,7 @@ class DictManager {
 
     private func replaceTextWithVars(_ text: String) -> String {
         let date = Date()
-        let arr = DictManager.dateFormatter.string(from: date).split(separator: " ")
+        let arr = DictReader.dateFormatter.string(from: date).split(separator: " ")
         let vars: [String: String] = [
             "{yyyy}": String(arr[0]),
             "{MM}": String(arr[1]),
@@ -246,11 +154,15 @@ class DictManager {
             label: "临时英文(空格输出半角符号,连敲;键两下输出全角符号)")]
     }
 
-    func getCandidates(query: String = String(), page: Int = 1) -> (candidates: [Candidate], hasNext: Bool) {
+    func getCandidates(
+        query: String,
+        page: Int,
+        tempEnTrigger: Character
+    ) -> (candidates: [Candidate], hasNext: Bool) {
         if query.count <= 0 {
             return ([], false)
         }
-        if query.first == tempEnTriggerPunctuation {
+        if query.first == tempEnTrigger {
             return (candidates: punctuationCandidates(query: query), hasNext: false)
         }
 
@@ -315,87 +227,6 @@ class DictManager {
         return (candidates, hasNext: allCount > count)
     }
 
-    func setCandidateToFirst(query: String, candidate: Candidate) {
-        let newCandidate = Candidate(code: query, text: candidate.text, type: CandidateType.user)
-        _ = prependCandidate(candidate: newCandidate)
-        NotificationQueue.default.enqueue(Notification(name: DictManager.userDictUpdated), postingStyle: .whenIdle)
-    }
-
-    func prependCandidate(candidate: Candidate) -> Bool {
-        // 插入用户词时自动从原始词库复制拆字/拼音，确保用户词也支持反查提示
-        let sql = """
-            insert into wb_py_dict(id, wbcode, text, type, query, spell, pinyin)
-            values (
-                (select MIN(id) - 1 from wb_py_dict), :code, :text, :type, :code,
-                (select spell from wb_py_dict where text = :text and spell is not null limit 1),
-                (select pinyin from wb_py_dict where text = :text and pinyin is not null limit 1)
-            );
-        """
-        var insertStatement: OpaquePointer?
-        if sqlite3_prepare_v2(database, sql, -1, &insertStatement, nil) == SQLITE_OK {
-            sqlite3_bind_text(insertStatement,
-                sqlite3_bind_parameter_index(insertStatement, ":code"),
-                              candidate.code, -1, SQLITE_TRANSIENT)
-            sqlite3_bind_text(insertStatement,
-                              sqlite3_bind_parameter_index(insertStatement, ":text"),
-                              candidate.text, -1, SQLITE_TRANSIENT)
-            sqlite3_bind_text(insertStatement,
-                              sqlite3_bind_parameter_index(insertStatement, ":type"),
-                              CandidateType.user.rawValue, -1, SQLITE_TRANSIENT)
-            if sqlite3_step(insertStatement) == SQLITE_DONE {
-                sqlite3_finalize(insertStatement)
-                insertStatement = nil
-                // 如果没有从码表复制到拆字数据（码表中无此词），按组词规则生成
-                fillCompoundGlyphs(for: candidate.text)
-                return true
-            }
-        }
-        sqlite3_finalize(insertStatement)
-        insertStatement = nil
-        let errMsg = database != nil ? String(cString: sqlite3_errmsg(database)) : "nil"
-        FireLog.dict.error("prependCandidate errmsg: \(errMsg, privacy: .public)")
-        return false
-    }
-
-    func deleteCandidate(_ candidate: Candidate) {
-        FireLog.dict.debug("deleteCandidate \(candidate.text) code=\(candidate.code) type=\(candidate.type.rawValue, privacy: .public)")
-        // 删除策略改为"屏蔽"而非直接删除，防止重建索引后被删除的词典词重新出现
-        // 用户词：删除 type='user' 记录清除排序调整，再插入 blocked 屏蔽原词典词
-        // 词典词：直接插入 blocked 屏蔽
-        if candidate.type == .user {
-            let sql = "delete from wb_py_dict where text = :text and type = :type"
-            var stmt: OpaquePointer?
-            if sqlite3_prepare_v2(database, sql, -1, &stmt, nil) == SQLITE_OK {
-                sqlite3_bind_text(stmt, sqlite3_bind_parameter_index(stmt, ":text"),
-                                  candidate.text, -1, SQLITE_TRANSIENT)
-                sqlite3_bind_text(stmt, sqlite3_bind_parameter_index(stmt, ":type"),
-                                  CandidateType.user.rawValue, -1, SQLITE_TRANSIENT)
-                if sqlite3_step(stmt) != SQLITE_DONE {
-                    FireLog.dict.error("delete user failed: \(String(cString: sqlite3_errmsg(self.database)), privacy: .public)")
-                }
-            }
-            sqlite3_finalize(stmt)
-        }
-        let blockSql = "insert or ignore into blocked_words(text) values (?)"
-        var blockStmt: OpaquePointer?
-        if sqlite3_prepare_v2(database, blockSql, -1, &blockStmt, nil) == SQLITE_OK {
-            sqlite3_bind_text(blockStmt, 1, candidate.text, -1, SQLITE_TRANSIENT)
-            if sqlite3_step(blockStmt) != SQLITE_DONE {
-                FireLog.dict.error("insert blocked_words failed: \(String(cString: sqlite3_errmsg(self.database)), privacy: .public)")
-            }
-        }
-        sqlite3_finalize(blockStmt)
-        NotificationQueue.default.enqueue(Notification(name: DictManager.userDictUpdated), postingStyle: .whenIdle)
-    }
-
-    /// 为多字词生成组合拆字并写入数据库（按五笔词组取码规则：2字各取前2码、3字前二字首码+末字前2码、≥4字前三字首码+末字首码）
-    /// - Parameters:
-    ///   - text: 候选词文本
-    ///   - skipIfExists: 若该列已有拆字数据则跳过（批量导入时使用）
-    private func fillCompoundGlyphs(for text: String, skipIfExists: Bool = false) {
-        DictGlyphFill.fillCompoundGlyphs(db: database, text: text, skipIfExists: skipIfExists)
-    }
-
     // 查询单个汉字的五笔全码(按长度降序取全码，避免拿到一码简码导致首根不全)
     func getCharFullWubiCode(_ char: String) -> String? {
         let sql = """
@@ -439,64 +270,6 @@ class DictManager {
             return prefix(fullCodes[0], 1) + prefix(fullCodes[1], 1)
                 + prefix(fullCodes[2], 1) + prefix(fullCodes[fullCodes.count - 1], 1)
         }
-    }
-
-    /// 批量插入用户词，使用参数化查询防止 SQL 注入
-    func prependCandidates(candidates: [Candidate]) {
-        if candidates.count <= 0 {
-            return
-        }
-        // 2.1 先获取最小id
-        let minId = getMinIdFromDictTable()
-        // 2.2 逐条插入（参数化查询，避免字符串拼接）
-        for (n, candidate) in candidates.enumerated() {
-            let id = minId - candidates.count + n
-            let sql = """
-                insert into wb_py_dict(id, wbcode, text, type, query, spell, pinyin)
-                values(?, ?, ?, ?, ?,
-                    (select spell from wb_py_dict where text = ? and spell is not null limit 1),
-                    (select pinyin from wb_py_dict where text = ? and pinyin is not null limit 1))
-            """
-            var stmt: OpaquePointer?
-            guard sqlite3_prepare_v2(database, sql, -1, &stmt, nil) == SQLITE_OK else { continue }
-            sqlite3_bind_int(stmt, 1, Int32(id))
-            sqlite3_bind_text(stmt, 2, candidate.code, -1, SQLITE_TRANSIENT)
-            sqlite3_bind_text(stmt, 3, candidate.text, -1, SQLITE_TRANSIENT)
-            sqlite3_bind_text(stmt, 4, candidate.type.rawValue, -1, SQLITE_TRANSIENT)
-            sqlite3_bind_text(stmt, 5, candidate.code, -1, SQLITE_TRANSIENT)
-            sqlite3_bind_text(stmt, 6, candidate.text, -1, SQLITE_TRANSIENT)
-            sqlite3_bind_text(stmt, 7, candidate.text, -1, SQLITE_TRANSIENT)
-            sqlite3_step(stmt)
-            sqlite3_finalize(stmt)
-        }
-        // 为没有拆字数据的多字词生成组合拆字
-        if candidates.count > 1 {
-            for candidate in candidates where candidate.text.count > 1 {
-                fillCompoundGlyphs(for: candidate.text, skipIfExists: true)
-            }
-        }
-    }
-
-    func updateUserDict(_ dictContent: String) {
-        // 使用事务确保删除+插入原子性，防止数据丢失
-        sqlite3_exec(database, "BEGIN TRANSACTION", nil, nil, nil)
-        // 1. 先删除之前的用户词库（type 为枚举常量，安全）
-        sqlite3_exec(database, "delete from wb_py_dict where type = '\(CandidateType.user.rawValue)'", nil, nil, nil)
-        // 2. 添加用户词库
-        let lines = dictContent.split(whereSeparator: \.isNewline)
-        FireLog.dict.debug("updateUserDict: \(lines.count, privacy: .public) lines")
-        let candidates = lines.map { (line) -> [Candidate] in
-            let parts = parseQuoteAware(line: line)
-            guard parts.count >= 2 else { return [] }
-            let code = String(parts[0])
-            let candidateTexts = parts[1...]
-            return candidateTexts.map { Candidate(code: code, text: String($0), type: CandidateType.user) }
-        }.reduce([] as [Candidate]) { partialResult, cur in
-            partialResult + cur
-        }
-        prependCandidates(candidates: candidates)
-        sqlite3_exec(database, "COMMIT", nil, nil, nil)
-        NotificationQueue.default.enqueue(Notification(name: DictManager.userDictUpdated), postingStyle: .whenIdle)
     }
 
     func getUserCandidates() -> [Candidate] {
@@ -549,14 +322,6 @@ class DictManager {
             words.append(String(cString: cstr))
         }
         return words
-    }
-
-    // 取消屏蔽：删除 type='blocked' 记录
-    func unblockWord(_ text: String) {
-        let sql = "delete from blocked_words where text = ?"
-        sqliteQuery(database, sql, bind: { stmt in
-            sqlite3_bind_text(stmt, 1, text, -1, SQLITE_TRANSIENT)
-        }) { _ in }
     }
 
     // 检查某个词是否已被屏蔽（用于组词时判断）
